refactor: moves integration page into domain
